LayoutTests/fast/events/touch/ios/block-without-overflow-scroll-and-passive-observer-on-block-scrolling-state-expected.txt

Verify that the block scrolling on overflow has touch event region with synchronous dispatch associated with it even with asynchronous listeners.

(Frame scrolling node
  (scrollable area size 800 600)
  (contents size 800 600)
  (synchronous event dispatch region for event touchend
    at (8,8) size 784x200)
  (synchronous event dispatch region for event touchstart
    at (8,8) size 784x200)
  (synchronous event dispatch region for event touchforcechange
    at (8,8) size 784x200)
  (synchronous event dispatch region for event touchmove
    at (8,8) size 784x200)
)
